+++ core/docs/changelog.md
@@ -3,6 +3,9 @@
 ### Fixes
 * Fixed Upload Files bug.
 
+### Changes
+* Added `EnclaveContext.UploadFiles` for uploading files to the API Container.
+
 # 1.43.6
 ### Features
 * Add `EnclaveContext.StoreFilesFromService` for copy files from a user service